requirements.txt

# Solo Git Dependencies

# Core dependencies
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0

# Git operations (Phase 1)
gitpython>=3.1.40

# Docker operations (Phase 1)
docker>=7.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Type stubs
types-PyYAML>=6.0.0
types-requests>=2.31.0

